OAuth overhaul, part 1

This is as much of the OAuth overhaul that will make it in to v0.5.0.
(I also want to close issue #235 before cutting v0.5.0, but otherwise
expect v0.5.0 to be this PR. Maybe #227 too.).

Closes #2. Closes #203. Closes #238.

### Developer workflow changes

This moves a lot of the OAuth protocol stuff out of apro.git and in to
https://github.com/datawire/liboauth2 .  Because that is a private
repo, you'll need to configure `git` such that `go get` will be able
to fetch it.  That means sticking this in your `~/.config/git/config`:

	[url "git@github.com:datawire/liboauth2"]
		insteadOf = https://github.com/datawire/liboauth2

Or, I've been happy making that substitution for all GitHub repos:

	[url "git@github.com:"]
		pushInsteadOf = https://github.com/
		insteadOf = https://github.com/

CI has already been taught about that change and been given permission
to read the liboauth2 repo.

### Customer-facing changes

 - The cookie is now named
   `ambassador_session.{{filter_name}}.{{filter_namespace}}` instead
   of `access_token`
 - The cookie is now an opaque string instead of a JWT Access Token.
   It is encouraged that the cookie be used as an XSRF token.  The
   Access Token is still available in the injected `Authorization`
   header.
 - It will no longer consider a user-agent-provided `Authorization`
   header, it will only look at the cookie.
 - Redis is now always required to be configured.
 - Refresh Tokens are supported; they must be requested by listing
   `offline_access` in the `scopes` argument in the FilterPolicy.
 - Much better error reporting.
 - I don't think it's possible to trigger a redirect loop anymore.
   Maybe if the Authorization Server does something like indicate that
   an Access Token is valid for 0 seconds; we'll immediately try to
   get a new one?  That's the only possibility I can think of as to
   how a loop could happen now.
 - The `audience` field in the OAuth2 `Filter` is now ignored.  I
   expect it to make a come-back in 0.5.1 though.
 - `/callback` should not be vulnerable to XSRF attacks
 - the `openid` scope value is *always* included in the authorization
   request, even if it's not in your configured list of scope values
   to ask for
 - How permissions are verified is configured by the OAuth2 `Filter`
   field `accessTokenValidation`, which takes a string, either `auto`,
   `jwt`, or `userinfo`. Empty or unset is equivalent to `auto`.
   `auto` has it do JWT validation if the Access Token parses like a
   JWT *and* the signature checks out, and otherwise falls back to
   polling the OIDC-core UserInfo Endpoint to validate it.  JWT
   validation works like it always has.  `userinfo` validation is slow
   but robust; it should _always_ work.
 - The `APRO_PRIVATE_KEY_PATH` and `APRO_PUBLIC_KEY_PATH` env
   variables are no longer used.  Instead, the keypair is configured
   as a Kubernetes secret.  The secret name is configured with the
   `APRO_KEYPAIR_SECRET_NAME` and `APRO_KEYPAIR_SECRET_NAMESPACE`,
   which default to `ambassador-pro-keypair` in the
   `$AMBASSADOR_NAMESPACE` namespace.  I forget what the expected
   format of the secret is, I'll need to document that.  If the secret
   does not exist, it will be automatically created; for this to
   happen it will need permission to create secrets.  If the secret
   already exists (from a prior run, or because it was manually
   created), the create permission is not needed.

## Testing

The test suite now verifies that we function properly with:
- Auth0
- Azure AD
- Google
- Keycloak
- Okta
- UAA

Google sometimes asks for a Captcha; if it does, then the Google test
is skipped, so not noticing a regression with Google is more likely
than with the other IDPs.

+++ docker/amb-sidecar/run.sh
@@ -5,6 +5,15 @@ export GRPC_PORT=${GRPC_PORT:-8501} # RLS gRPC
 export DEBUG_PORT=${DEBUG_PORT:-8502} # RLS debug (HTTP)
 export PORT=${PORT:-8503} # RLS HTTP ???
 
+if test -z "$REDIS_URL"; then
+	echo 'Error: ${REDIS_URL} is not set; not starting'
+	exit 1
+fi
+if test -z "$REDIS_SOCKET_TYPE"; then
+	echo 'Error: ${REDIS_SOCKET_TYPE} is not set; not starting'
+	exit 1
+fi
+
 exe="${BASH_SOURCE[0]%/*}/amb-sidecar"
 trap 'jobs -p | xargs -r kill --' INT
 
@@ -21,14 +30,7 @@ mkdir -p ${RUN_DIR}
 export RLS_RUNTIME_DIR=${RUN_DIR}/config
 
 # Launch each of the worker processes
-if test -z "$REDIS_URL"; then
-	echo 'Warning: ${REDIS_URL} is not set; not starting ratelimit service'
-else
-	# Setting the PORT is important only because the default PORT
-	# is 8080, which would clash with non-root-Ambassador when
-	# running as a sidecar.
-	launch USE_STATSD=false RUNTIME_ROOT=${RUN_DIR}/config RUNTIME_SUBDIRECTORY=config "$exe" ratelimit
-fi
+launch USE_STATSD=false RUNTIME_ROOT=${RUN_DIR}/config RUNTIME_SUBDIRECTORY=config "$exe" ratelimit
 launch "$exe" main
 
 # Wait for one of them to quit, then kill the others
